requirements.txt
fastembed==0.3.3
flashrank==0.2.8
langchain-community==0.2.6
langchain-experimental==0.0.62
langchain-groq==0.1.6
langchain-qdrant==0.1.1
langchain==0.2.6
python-dotenv==1.0.1
qdrant-client==1.10.1
streamlit==1.36.0
pypdfium2==4.30.0
PyMuPDF
langchain-ollama

# numpy
# pandas
# torch
# spacy
# transformers
# scikit-learn
# google-generativeai